cleaned up inapp manager to avoid exposing unwanted methods
